refactor: split cli and exporter helpers
